docs/requirements.txt
sphinx==8.1.3
furo==2024.8.6
sphinx-copybutton==0.5.2
sphinx-fortran==1.1.1
sphinx-tabs==3.4.7
breathe==4.35.0
six==1.17.0
ford==7.0.9
markdown_checklist==0.4.4
enum-tools[sphinx]
dtfft-openmpi
--extra-index-url https://test.pypi.org/simple/
